# perf(table): Reduce calls to updateStickyColumnStyles