Tested with the dev app, clicking "add table" in the "Table with sticky headers, footers and columns" example.

Before:
![image](https://user-images.githubusercontent.com/923677/85492960-06b4d600-b5a4-11ea-8df5-2b2b12441abf.png)
100ms scripting + 59ms rendering

After:
![image](https://user-images.githubusercontent.com/923677/85493056-3368ed80-b5a4-11ea-9663-d65e01f1f13c.png)
75ms scripting + 39ms rendering

The effect is greater in more complex tables. More optimization PRs to come.