@andrewseguin 2020-06-24
Awesome thanks for catching this optimization - looks good to me, just small nits on readability